Refactor product details hooks and fix upsert id
