+++ packages/plugin-commerce-dashboard/src/modules/products/hooks/use-product-details.ts
@@ -9,221 +9,26 @@ import {
 } from "@kitejs-cms/dashboard-core";
 import type {
   ProductResponseDetailsModel,
-  ProductTranslationModel,
   ProductUpsertModel,
-  ProductSeoModel,
-  ProductPriceModel,
   ProductVariantModel,
+  ProductSeoModel,
 } from "@kitejs-cms/plugin-commerce-api";
+import { useProductOptions } from "./use-product-options";
+import { useProductVariants } from "./use-product-variants";
+import {
+  arraysEqual,
+  extractAssetId,
+  formatOptionDisplayName,
+  generateSlug,
+  hydrateProductDetails,
+  mapAssetIdsToSources,
+  normalizeThumbnailIdentifier,
+  slugifyOptionHandle,
+  getTranslationCandidate,
+} from "./product-details.utils";
+import type { ProductDetailsState } from "./product-details.types";
 import type { MediaSource } from "./use-product-media";
-import type {
-  VariantFieldErrors,
-  VariantGalleryOption,
-} from "../components/variants-section";
-
-const arraysEqual = (a: string[], b: string[]) =>
-  a.length === b.length && a.every((value, index) => value === b[index]);
-
-const generateSlug = (title: string) =>
-  title
-    .toLowerCase()
-    .trim()
-    .replace(/[\s\W-]+/g, "-");
-
-const slugifyOptionHandle = (value: string) =>
-  value
-    .normalize("NFKD")
-    .replace(/[\u0300-\u036f]/g, "")
-    .toLowerCase()
-    .trim()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-
-const formatOptionDisplayName = (value: string) =>
-  value
-    .split(/[-_\s]+/)
-    .filter(Boolean)
-    .map((segment) => segment[0]?.toUpperCase() + segment.slice(1))
-    .join(" ");
-
-const getTranslationCandidate = (
-  translations: Record<string, ProductTranslationModel>,
-  lang?: string
-) => {
-  if (!lang) return null;
-  const translation = translations[lang];
-  if (translation?.title?.trim() && translation?.slug?.trim()) {
-    return { lang, translation } as const;
-  }
-  return null;
-};
-
-type ProductDetailsState = Omit<
-  ProductResponseDetailsModel,
-  "gallery" | "thumbnail"
-> & {
-  gallery: MediaSource[];
-  thumbnail?: string | null;
-};
-
-const extractAssetId = (source: MediaSource | undefined | null): string | null => {
-  if (!source) return null;
-  if (typeof source === "string") return source;
-  return (
-    source.assetId ??
-    source.id ??
-    source._id ??
-    source.path ??
-    null
-  );
-};
-
-const mapAssetIdsToSources = (
-  assetIds: string[],
-  previous: MediaSource[]
-): MediaSource[] => {
-  if (!assetIds.length) return [];
-
-  const lookup = new Map<string, MediaSource>();
-  previous.forEach((entry) => {
-    const key = extractAssetId(entry);
-    if (key) {
-      lookup.set(key, entry);
-    }
-  });
-
-  return assetIds.map((id) => lookup.get(id) ?? id);
-};
-
-const isLikelyExternalUrl = (value: string) => /^https?:\/\//i.test(value);
-
-const getUrlPath = (value: string): string | null => {
-  if (!value) return null;
-
-  try {
-    const url = new URL(value);
-    return url.pathname;
-  } catch (error) {
-    const normalized = value.split("?")[0];
-    return normalized.startsWith("/") ? normalized : null;
-  }
-};
-
-const findGalleryAssetIdByUrl = (
-  gallery: MediaSource[],
-  url: string
-): string | null => {
-  const targetPath = getUrlPath(url);
-  if (!targetPath) return null;
-
-  for (const entry of gallery) {
-    if (!entry) continue;
-
-    if (typeof entry === "string") {
-      if (!isLikelyExternalUrl(entry)) continue;
-      const entryPath = getUrlPath(entry);
-      if (entryPath && entryPath === targetPath) {
-        // String entries don't provide a stable asset identifier, so skip.
-        continue;
-      }
-      continue;
-    }
-
-    const urlsToCompare: string[] = [];
-    if (typeof entry.url === "string") {
-      urlsToCompare.push(entry.url);
-    }
-    const previewUrl = (entry as { previewUrl?: string | null }).previewUrl;
-    if (typeof previewUrl === "string") {
-      urlsToCompare.push(previewUrl);
-    }
-
-    for (const candidate of urlsToCompare) {
-      const candidatePath = getUrlPath(candidate);
-      if (candidatePath && candidatePath === targetPath) {
-        const assetId = extractAssetId(entry);
-        if (assetId) return assetId;
-      }
-    }
-
-    const entryPath = (entry as { path?: string | null }).path;
-    if (typeof entryPath === "string" && targetPath.endsWith(entryPath)) {
-      const assetId = extractAssetId(entry);
-      if (assetId) return assetId;
-    }
-  }
-
-  return null;
-};
-
-const normalizeThumbnailIdentifier = (
-  thumbnail: string | null | undefined,
-  gallery: MediaSource[]
-): string | null => {
-  if (!thumbnail) return null;
-  if (!isLikelyExternalUrl(thumbnail)) return thumbnail;
-
-  return findGalleryAssetIdByUrl(gallery, thumbnail);
-};
-
-const hydrateProductDetails = (
-  product: ProductResponseDetailsModel
-): ProductDetailsState => {
-  const gallery = (product.gallery ?? []) as MediaSource[];
-  const normalizedThumbnail = normalizeThumbnailIdentifier(
-    product.thumbnail ?? null,
-    gallery
-  );
-
-  const normalizedOptions = (product.options ?? [])
-    .map((option, index) => {
-      const trimmedName = option?.name?.trim() ?? "";
-      const slug = trimmedName
-        ? slugifyOptionHandle(trimmedName)
-        : slugifyOptionHandle(option?.displayName ?? "");
-
-      const displayName = option?.displayName?.trim();
-
-      const values = Array.isArray(option?.values)
-        ? option.values
-            .map((value) => value?.trim())
-            .filter((value): value is string => Boolean(value))
-        : [];
-
-      return {
-        name: slug || `option-${index + 1}`,
-        displayName:
-          displayName && displayName.length > 0
-            ? displayName
-            : trimmedName
-              ? formatOptionDisplayName(trimmedName)
-              : `Option ${index + 1}`,
-        values,
-        position:
-          typeof option?.position === "number" ? option.position : index,
-      } satisfies ProductDetailsState["options"][number];
-    })
-    .sort((a, b) => a.position - b.position);
-
-  const normalizedVariants = (product.variants ?? []).map((variant) => ({
-    ...variant,
-    optionName: variant.optionName?.trim() ?? undefined,
-    optionValue: variant.optionValue?.trim() ?? undefined,
-    gallery: Array.isArray(variant.gallery)
-      ? variant.gallery.filter(
-          (id): id is string => typeof id === "string" && Boolean(id)
-        )
-      : [],
-  }));
-
-  return {
-    ...product,
-    gallery,
-    thumbnail: normalizedThumbnail ?? product.thumbnail ?? null,
-    options: normalizedOptions,
-    variants: normalizedVariants,
-  };
-};
+import type { VariantFieldErrors } from "../components/variants-section";
 
 const findTranslationForUpsert = (
   product: ProductDetailsState,
@@ -749,8 +554,14 @@ export function useProductDetails() {
         (localData.gallery ?? []) as MediaSource[]
       );
 
+      const upsertId = localData?.id?.trim()
+        ? localData.id.trim()
+        : id && id !== "create"
+          ? id
+          : undefined;
+
       const body: ProductUpsertModel = {
-        id: id && id !== "create" ? localData.id : undefined,
+        id: upsertId,
         language: activeLang,
         status: localData.status,
         slug: localData.translations[activeLang].slug,
@@ -934,7 +745,8 @@ export function useProductDetails() {
       const { lang, translation } = translationEntry;
 
       const body: ProductUpsertModel = {
-        id: updatedData.id,
+        id:
+          updatedData.id?.trim() || (id && id !== "create" ? id : undefined),
         language: lang,
         status: updatedData.status,
         slug: translation.slug,
@@ -965,620 +777,34 @@ export function useProductDetails() {
         throw error;
       }
     },
-    [activeLang, defaultLang, fetchData, localData, t]
+    [activeLang, defaultLang, fetchData, id, localData, t]
   );
 
-  const onVariantChange = useCallback(
-    (
-      index: number,
-      field:
-        | "title"
-        | "sku"
-        | "barcode"
-        | "inventoryQuantity"
-        | "allowBackorder"
-        | "optionName"
-        | "optionValue",
-      value: string | number | boolean | undefined
-    ) => {
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const variants = [...(prev.variants ?? [])];
-        const current = variants[index];
-        if (!current) return prev;
-
-        const updated = { ...current } as typeof current;
-
-        switch (field) {
-          case "title":
-          case "sku":
-          case "barcode":
-            updated[field] = (value as string) ?? "";
-            break;
-          case "inventoryQuantity": {
-            const numericValue =
-              typeof value === "number"
-                ? value
-                : value === undefined
-                  ? undefined
-                  : Number(value);
-
-            updated.inventoryQuantity =
-              numericValue === undefined || Number.isNaN(numericValue)
-                ? 0
-                : Math.max(0, Math.trunc(numericValue));
-            break;
-          }
-          case "allowBackorder":
-            updated.allowBackorder = Boolean(value);
-            break;
-          case "optionName": {
-            const rawValue = typeof value === "string" ? value : "";
-            const normalizedValue = rawValue
-              ? slugifyOptionHandle(rawValue)
-              : "";
-            const availableOptions = prev.options ?? [];
-            const matchedOption = availableOptions.find(
-              (option) => option.name === rawValue || option.name === normalizedValue
-            );
-
-            if (matchedOption) {
-              updated.optionName = matchedOption.name;
-              if (
-                matchedOption.values?.length &&
-                matchedOption.values.includes(updated.optionValue ?? "")
-              ) {
-                // Keep current option value if it's still valid
-                updated.optionValue = updated.optionValue;
-              } else {
-                updated.optionValue = matchedOption.values?.[0];
-              }
-            } else {
-              updated.optionName =
-                normalizedValue && normalizedValue.length > 0
-                  ? normalizedValue
-                  : undefined;
-              // Preserve existing option value when working with custom handles
-            }
-            break;
-          }
-          case "optionValue": {
-            if (typeof value === "string") {
-              const trimmed = value.trim();
-              updated.optionValue = trimmed.length > 0 ? trimmed : undefined;
-            } else {
-              updated.optionValue = undefined;
-            }
-            break;
-          }
-        }
-
-        variants[index] = updated;
-        return { ...prev, variants };
-      });
-      setVariantErrors((prev) => {
-        if (!prev.length) return prev;
-        const current = prev[index];
-        if (!current) return prev;
-        const fieldKey = field as keyof VariantFieldErrors;
-        if (!current[fieldKey]) {
-          return prev;
-        }
-
-        const next = [...prev];
-        const updatedErrors = { ...current } as VariantFieldErrors;
-        delete updatedErrors[fieldKey];
-        next[index] = Object.keys(updatedErrors).length ? updatedErrors : {};
-        return next;
-      });
-      setHasChanges(true);
-    },
-    []
-  );
-
-  const onVariantPriceChange = useCallback(
-    (index: number, field: keyof ProductPriceModel, value: string) => {
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const variants = [...(prev.variants ?? [])];
-        const current = variants[index];
-        if (!current) return prev;
-
-        const prices = [...(current.prices ?? [])];
-        const primary = {
-          currencyCode: defaultCurrency,
-          amount: 0,
-          compareAtAmount: undefined as number | undefined,
-          ...prices[0],
-        };
-
-        if (field === "currencyCode") {
-          primary.currencyCode = value
-            ? value.trim().toUpperCase().slice(0, 3)
-            : defaultCurrency;
-        }
-
-        if (field === "amount") {
-          const parsed = Number.parseFloat(value);
-          primary.amount = Number.isNaN(parsed) ? 0 : Math.max(0, parsed);
-        }
-
-        if (field === "compareAtAmount") {
-          const parsed = Number.parseFloat(value);
-          primary.compareAtAmount = Number.isNaN(parsed)
-            ? undefined
-            : Math.max(0, parsed);
-        }
-
-        prices[0] = primary;
-        variants[index] = { ...current, prices };
-        return { ...prev, variants };
-      });
-      setHasChanges(true);
-    },
-    [defaultCurrency]
-  );
-
-  const onVariantGalleryChange = useCallback(
-    (index: number, gallery: string[]) => {
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const variants = [...(prev.variants ?? [])];
-        const current = variants[index];
-        if (!current) return prev;
-
-        variants[index] = {
-          ...current,
-          gallery,
-        };
-
-        return { ...prev, variants };
-      });
-
-      setVariantErrors((prev) => {
-        if (!prev.length) return prev;
-        const current = prev[index];
-        if (!current?.gallery) return prev;
-        const next = [...prev];
-        const updated = { ...current } as VariantFieldErrors;
-        delete updated.gallery;
-        next[index] = Object.keys(updated).length ? updated : {};
-        return next;
-      });
-
-      setHasChanges(true);
-    },
-    []
-  );
-
-  const onAddOption = useCallback(() => {
-    setLocalData((prev) => {
-      if (!prev) return prev;
-
-      const existingOptions = prev.options ?? [];
-      const position = existingOptions.length;
-      const defaultLabel = t("products.details.options.defaultName", {
-        index: position + 1,
-      });
-      const baseHandle = slugifyOptionHandle(defaultLabel) || `option-${position + 1}`;
-
-      let handle = baseHandle;
-      let attempt = 1;
-      const existingHandles = new Set(existingOptions.map((option) => option.name));
-      while (existingHandles.has(handle)) {
-        handle = `${baseHandle}-${++attempt}`;
-      }
-
-      const option: ProductDetailsState["options"][number] = {
-        name: handle,
-        displayName: defaultLabel,
-        values: [],
-        position,
-      };
-
-      return {
-        ...prev,
-        options: [...existingOptions, option],
-      };
-    });
-    setHasChanges(true);
-  }, [t]);
-
-  const onRemoveOption = useCallback(
-    (index: number) => {
-      const clearedIndexes = new Set<number>();
-      const missingOptionIndexes = new Set<number>();
-      const missingValueIndexes = new Set<number>();
-
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const options = [...(prev.options ?? [])];
-        const target = options[index];
-        if (!target) return prev;
-
-        options.splice(index, 1);
-
-        const normalizedOptions = options.map((option, optionIndex) => ({
-          ...option,
-          position: optionIndex,
-        }));
-
-        const fallbackOption = normalizedOptions[0];
-
-        const variants = (prev.variants ?? []).map((variant, variantIndex) => {
-          if (variant.optionName === target.name) {
-            if (fallbackOption?.name) {
-              const hasExistingValue = fallbackOption.values?.includes(
-                variant.optionValue ?? ""
-              );
-              const fallbackValue = hasExistingValue
-                ? variant.optionValue
-                : fallbackOption.values?.[0];
-
-              if (fallbackValue) {
-                clearedIndexes.add(variantIndex);
-              } else {
-                missingValueIndexes.add(variantIndex);
-              }
-
-              return {
-                ...variant,
-                optionName: fallbackOption.name,
-                optionValue: fallbackValue,
-              };
-            }
-
-            missingOptionIndexes.add(variantIndex);
-            missingValueIndexes.add(variantIndex);
-            return {
-              ...variant,
-              optionName: undefined,
-              optionValue: undefined,
-            };
-          }
-
-          return variant;
-        });
-
-        return {
-          ...prev,
-          options: normalizedOptions,
-          variants,
-        };
-      });
-
-      if (
-        clearedIndexes.size ||
-        missingOptionIndexes.size ||
-        missingValueIndexes.size
-      ) {
-        setVariantErrors((prev) => {
-          const next = [...prev];
-          clearedIndexes.forEach((idx) => {
-            while (next.length <= idx) {
-              next.push({});
-            }
-            const current = next[idx];
-            if (!current) return;
-            const updated = { ...current } as VariantFieldErrors;
-            delete updated.optionName;
-            delete updated.optionValue;
-            next[idx] = Object.keys(updated).length ? updated : {};
-          });
-
-          missingOptionIndexes.forEach((idx) => {
-            while (next.length <= idx) {
-              next.push({});
-            }
-            const current = next[idx] ?? {};
-            next[idx] = {
-              ...current,
-              optionName: t(
-                "products.errors.variantOptionRequired",
-                "Select an option"
-              ),
-              optionValue: t(
-                "products.errors.variantOptionValueRequired",
-                "Select a value"
-              ),
-            };
-          });
-
-          missingValueIndexes.forEach((idx) => {
-            while (next.length <= idx) {
-              next.push({});
-            }
-            if (missingOptionIndexes.has(idx)) return;
-            const current = next[idx] ?? {};
-            next[idx] = {
-              ...current,
-              optionValue: t(
-                "products.errors.variantOptionValueRequired",
-                "Select a value"
-              ),
-            };
-          });
-
-          return next;
-        });
-      }
-
-      setHasChanges(true);
-    },
-    [t]
-  );
-
-  const onOptionChange = useCallback(
-    (
-      index: number,
-      field: "name" | "displayName",
-      value: string
-    ) => {
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const options = [...(prev.options ?? [])];
-        const target = options[index];
-        if (!target) return prev;
-
-        const updatedOption = { ...target } as ProductDetailsState["options"][number];
-        if (field === "displayName") {
-          updatedOption.displayName = value;
-          const currentSlug = slugifyOptionHandle(target.displayName ?? "");
-          const currentHandle = target.name ?? "";
-          const nextSlug = slugifyOptionHandle(value);
-
-          if (!currentHandle || currentHandle === currentSlug) {
-            updatedOption.name = nextSlug || currentHandle;
-          }
-        } else {
-          const nextHandle = slugifyOptionHandle(value);
-          updatedOption.name = nextHandle;
-
-          const variants = (prev.variants ?? []).map((variant) => {
-            if (variant.optionName === target.name) {
-              return {
-                ...variant,
-                optionName: nextHandle,
-              };
-            }
-            return variant;
-          });
-
-          options[index] = {
-            ...updatedOption,
-            position: target.position,
-          };
-
-          return {
-            ...prev,
-            options,
-            variants,
-          };
-        }
-
-        options[index] = {
-          ...updatedOption,
-          position: target.position,
-        };
-
-        return {
-          ...prev,
-          options,
-        };
-      });
-
-      setHasChanges(true);
-    },
-    []
-  );
-
-  const onOptionValueAdd = useCallback(
-    (index: number, value: string) => {
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const options = [...(prev.options ?? [])];
-        const target = options[index];
-        if (!target) return prev;
-
-        const trimmedValue = value.trim();
-        if (!trimmedValue) return prev;
-
-        const existingValues = new Set(target.values ?? []);
-        if (existingValues.has(trimmedValue)) {
-          return prev;
-        }
-
-        const nextValues = [...(target.values ?? []), trimmedValue];
-
-        options[index] = {
-          ...target,
-          values: nextValues,
-        };
-
-        return {
-          ...prev,
-          options,
-        };
-      });
-
-      setHasChanges(true);
-    },
-    []
-  );
-
-  const onOptionValueRemove = useCallback(
-    (index: number, value: string) => {
-      const missingValueIndexes = new Set<number>();
-
-      setLocalData((prev) => {
-        if (!prev) return prev;
-        const options = [...(prev.options ?? [])];
-        const target = options[index];
-        if (!target) return prev;
-
-        const filteredValues = (target.values ?? []).filter(
-          (entry) => entry !== value
-        );
-
-        options[index] = {
-          ...target,
-          values: filteredValues,
-        };
-
-        const variants = (prev.variants ?? []).map((variant, variantIndex) => {
-          if (variant.optionName === target.name && variant.optionValue === value) {
-            const fallbackValue = filteredValues[0];
-            if (fallbackValue) {
-              return {
-                ...variant,
-                optionValue: fallbackValue,
-              };
-            }
-            missingValueIndexes.add(variantIndex);
-            return {
-              ...variant,
-              optionValue: undefined,
-            };
-          }
-          return variant;
-        });
-
-        return {
-          ...prev,
-          options,
-          variants,
-        };
-      });
-
-      if (missingValueIndexes.size) {
-        setVariantErrors((prev) => {
-          const next = [...prev];
-          missingValueIndexes.forEach((idx) => {
-            while (next.length <= idx) {
-              next.push({});
-            }
-            const current = next[idx] ?? {};
-            next[idx] = {
-              ...current,
-              optionValue: t(
-                "products.errors.variantOptionValueRequired",
-                "Select a value"
-              ),
-            };
-          });
-
-          return next;
-        });
-      }
-
-      setHasChanges(true);
-    },
-    [t]
-  );
-
-  const onAddVariant = useCallback(() => {
-    let didAdd = false;
-    setLocalData((prev) => {
-      if (!prev) return prev;
-
-      const primaryOption = prev.options?.[0];
-      const optionHandle = primaryOption?.name?.trim() || "default";
-      const initialOptionValue =
-        primaryOption?.values?.[0] ?? `${optionHandle}-${Date.now()}`;
-
-      const variant: ProductDetailsState["variants"][number] = {
-        id: undefined,
-        title: "",
-        sku: "",
-        barcode: "",
-        prices: [
-          {
-            currencyCode: defaultCurrency,
-            amount: 0,
-          },
-        ],
-        inventoryQuantity: 0,
-        allowBackorder: false,
-        gallery: [],
-        optionName: optionHandle,
-        optionValue: initialOptionValue,
-      } as ProductDetailsState["variants"][number];
-
-      didAdd = true;
-
-      return {
-        ...prev,
-        variants: [...(prev.variants ?? []), variant],
-      };
-    });
-    if (didAdd) {
-      setVariantErrors((prev) => [...prev, {}]);
-    }
-    setHasChanges(true);
-  }, [defaultCurrency]);
-
-  const onRemoveVariant = useCallback((index: number) => {
-    let didRemove = false;
-    setLocalData((prev) => {
-      if (!prev) return prev;
-      const variants = [...(prev.variants ?? [])];
-      if (!variants[index]) return prev;
-      variants.splice(index, 1);
-      didRemove = true;
-      return { ...prev, variants };
-    });
-    if (didRemove) {
-      setVariantErrors((prev) => {
-        if (!prev.length) return prev;
-        const next = [...prev];
-        next.splice(index, 1);
-        return next;
-      });
-    }
-    setHasChanges(true);
-  }, []);
-
-  const variantGalleryOptions = useMemo<VariantGalleryOption[]>(() => {
-    if (!localData) return [];
-
-    const languages = [activeLang, defaultLang].filter(
-      (lang): lang is string => Boolean(lang)
-    );
-
-    return (localData.gallery ?? [])
-      .map((entry) => {
-        const assetId = extractAssetId(entry);
-        if (!assetId) return null;
-
-        if (typeof entry === "string") {
-          return {
-            id: assetId,
-            label: assetId,
-            url: isLikelyExternalUrl(entry) ? entry : null,
-          } satisfies VariantGalleryOption;
-        }
-
-        const localizedLabel = languages.reduce<string | undefined>(
-          (acc, lang) => {
-            if (acc) return acc;
-            const title = entry.title?.[lang]?.trim();
-            if (title) return title;
-            const alt = entry.alt?.[lang]?.trim();
-            if (alt) return alt;
-            return acc;
-          },
-          undefined
-        );
-
-        return {
-          id: assetId,
-          label:
-            localizedLabel ?? entry.name ?? entry.path ?? assetId ?? "",
-          url: entry.url ?? null,
-        } satisfies VariantGalleryOption;
-      })
-      .filter((option): option is VariantGalleryOption => option !== null);
-  }, [localData, activeLang, defaultLang]);
+  const { productOptions, onAddOption, onRemoveOption, onOptionChange, onOptionValueAdd, onOptionValueRemove } = useProductOptions({
+    localData,
+    setLocalData,
+    setHasChanges,
+    setVariantErrors,
+    t,
+  });
+
+  const {
+    variantGalleryOptions,
+    onVariantChange,
+    onVariantPriceChange,
+    onVariantGalleryChange,
+    onAddVariant,
+    onRemoveVariant,
+  } = useProductVariants({
+    localData,
+    setLocalData,
+    setHasChanges,
+    setVariantErrors,
+    defaultCurrency,
+    productOptions,
+    activeLang,
+    defaultLang,
+  });
 
   return {
     data: localData,
@@ -1611,6 +837,6 @@ export function useProductDetails() {
     onOptionValueRemove,
     variantErrors,
     variantGalleryOptions,
-    productOptions: localData?.options ?? [],
+    productOptions,
   };
 }
